feat:replaced 64

+++ gymnax/environments/misc/empty.py
@@ -186,10 +186,10 @@ class Empty(environment.Environment):
     def observation_space(self, params: EnvParams) -> spaces.Box:
         """Observation space of the environment."""
         if self.use_visual_obs:
-            return spaces.Box(0, 1, (16, 16, 2), jnp.float32)
+            return spaces.Box(0, 1, (16, 16, 2), jnp.float64)
         else:
             return spaces.Box(
-                jnp.min(self.coords), jnp.max(self.coords), (4,), jnp.float32
+                jnp.min(self.coords), jnp.max(self.coords), (4,), jnp.float64
             )
 
     def state_space(self, params: EnvParams) -> spaces.Dict:
@@ -200,13 +200,13 @@ class Empty(environment.Environment):
                     jnp.min(self.coords),
                     jnp.max(self.coords),
                     (2,),
-                    jnp.float32,
+                    jnp.float64,
                 ),
                 "goal": spaces.Box(
                     jnp.min(self.coords),
                     jnp.max(self.coords),
                     (2,),
-                    jnp.float32,
+                    jnp.float64,
                 ),
                 "time": spaces.Discrete(params.max_steps_in_episode),
             }
